Person info displayed

+++ src/main/java/model/CreditCard.java
@@ -36,7 +36,7 @@ public class CreditCard extends Product{
 		return debt;
 	}
 
-	public double getDayOfPayment() {
+	public int getDayOfPayment() {
 		return dayOfPayment;
 	}
